pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fluid-rower-monitor"
version = "0.1.0"
description = "Monitor and analyze rowing sessions from a Fluid Rower device via serial connection"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Kold"}
]
keywords = ["rowing", "fitness", "tracking", "serial", "monitor"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Sports",
]

dependencies = [
    "pyserial>=3.5",
    "pandas>=1.3.0",
    "polars>=0.18.0",
    "pyarrow>=10.0.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",
    "fastapi>=0.103,<1.0",
    "uvicorn[standard]>=0.23,<1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=3.0",
    "httpx>=0.27",
    "tox>=3.24",
    "flake8>=4.0",
    "black>=22.0",
    "tomli>=2.0.0; python_version < '3.11'",
]

[project.urls]
Homepage = "https://github.com/kold/fluid_rower_usb_monitor"
Repository = "https://github.com/kold/fluid_rower_usb_monitor.git"
Issues = "https://github.com/kold/fluid_rower_usb_monitor/issues"

[project.scripts]
fluid-rower-monitor = "fluid_rower_monitor.serial_conn:main"
fluid-rower-monitor-api = "fluid_rower_monitor.api.main:run"

# Tool configurations
[tool.black]
line-length = 120
target-version = ['py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 120
skip_gitignore = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
]

[tool.coverage.run]
source = ["fluid_rower_monitor"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    'if __name__ == .__main__.:',
    "if TYPE_CHECKING:",
    "@abstractmethod",
    "@abc.abstractmethod",
]
precision = 2
fail_under = 80

[tool.flake8]
max-line-length = 120
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".tox",
    ".venv",
    "venv",
    ".eggs",
    "*.egg",
]
ignore = ["W503", "E501"]

[tool.tox]
legacy_tox_ini = "tox.ini"
[tool.setuptools.packages.find]
where = ["."]
include = ["fluid_rower_monitor*"]
exclude = ["rowing_sessions*", "tests*", "htmlcov*"]